requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pytest>=7.4.0
python-dotenv>=1.0.0
